requirements.txt
# FastAPI and web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40

# MCP (Model Context Protocol)
mcp>=0.9.0

# Vector store
faiss-cpu>=1.7.4
numpy>=1.24.0

# Redis
redis>=5.0.0

# Git operations
gitpython>=3.1.40

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
